requirements.txt
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.2
python-dotenv>=0.21.0
numpy>=1.24.0,<3.0.0
